requirements.txt
requests
curl_cffi
playwright
beautifulsoup4
dateparser
python-dotenv
google-genai
